fixing warnings
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/branches/distributed@12477 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ exp/include/Galois/Runtime/Serialize.h
@@ -49,7 +49,7 @@ public:
 
   SerializeBuffer() {
     //reserve a header
-    for (int i = 0; i < sizeof(uintptr_t); ++i)
+    for (size_t i = 0; i < sizeof(uintptr_t); ++i)
       bufdata.push_back(0);
   }
